+++ src/Controller/Admin/SettingsController.php
@@ -5,33 +5,53 @@ namespace App\Controller\Admin;
 use App\Controller\SetupController;
 use App\Entity\Settings;
 use App\Entity\SidebarContent;
+use App\Entity\User;
+use App\Entity\UserSettings;
+use App\Repository\SettingsRepository;
+use App\Repository\SidebarContentRepository;
 use App\Utilities\LatLong;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\CurrentUser;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 #[IsGranted('ROLE_SUPER_ADMIN', statusCode: 403, exceptionCode: 10010)]
 class SettingsController extends AbstractController
 {
+    private SettingsRepository $settingsRepository;
+    private SidebarContentRepository $sidebarContentRepository;
+
     public function __construct(
         private TranslatorInterface $translator,
+        private EntityManagerInterface $entityManager,
     ) {
+        $this->settingsRepository = $this->entityManager->getRepository(Settings::class);
+        $this->sidebarContentRepository = $this->entityManager->getRepository(SidebarContent::class);
     }
 
     #[Route(path: '/admin/settings', name: 'app_admin_settings', methods: ['GET', 'POST'])]
     public function index(
         Request $request,
-        EntityManagerInterface $entityManager,
+        #[CurrentUser] ?User $user,
     ): Response {
-        $settingsRepository = $entityManager->getRepository(Settings::class);
-        $sidebarContentRepository = $entityManager->getRepository(SidebarContent::class);
-        $sidebarContent = $sidebarContentRepository->findOneBy([]);
+        $sidebarContent = $this->sidebarContentRepository->findOneBy([]);
         if (!$sidebarContent) {
             $sidebarContent = new SidebarContent();
+        }
+
+        $userSettingsRepository = $this->entityManager->getRepository(UserSettings::class);
+        $adminUserThemeSetting = $userSettingsRepository->findOneBy([
+            'name' => 'theme',
+            'user' => $user->getId(),
+        ]);
+
+        $adminUserTheme = null;
+        if ($adminUserThemeSetting) {
+            $adminUserTheme = $this->translator->trans('fields.theme.' . $adminUserThemeSetting->getValue());
         }
 
         if ('POST' === $request->getMethod()) {
@@ -44,6 +64,7 @@ class SettingsController extends AbstractController
 
                 return $this->render('admin/settings.html.twig', [
                     'languages' => SetupController::LANGUAGES,
+                    'adminUserTheme' => $adminUserTheme,
                     'sidebarContent' => ($sidebarContent->getContent()) ? $sidebarContent->getContent() : '',
                 ]);
             }
@@ -53,6 +74,7 @@ class SettingsController extends AbstractController
             if (!empty($fieldErrors)) {
                 return $this->render('admin/settings.html.twig', [
                     'languages' => SetupController::LANGUAGES,
+                    'adminUserTheme' => $adminUserTheme,
                     'errors' => $fieldErrors,
                     'values' => [
                         'habitatName' => $request->request->get('habitatName'),
@@ -61,73 +83,25 @@ class SettingsController extends AbstractController
                         'locationRadiusMeters' => $request->request->get('locationRadiusMeters'),
                         'locationZoom' => $request->request->get('locationZoom'),
                         'language' => $request->request->get('language'),
+                        'theme' => $request->request->get('theme'),
                         'sidebarContent' => $request->request->get('sidebarContent'),
                     ],
                 ]);
             }
 
-            $habitatNameSetting = $settingsRepository->getSettingByName('habitatName');
-            if (!$habitatNameSetting) {
-                $habitatNameSetting = new Settings();
-                $habitatNameSetting->setName('habitatName');
-            }
-            $habitatNameSetting->setValue($request->request->get('habitatName'));
-
-            $entityManager->persist($habitatNameSetting);
-
-            $locationRadiusSetting = $settingsRepository->getSettingByName('locationRadiusMeters');
-            if (!$locationRadiusSetting) {
-                $locationRadiusSetting = new Settings();
-                $locationRadiusSetting->setName('locationRadiusMeters');
-            }
-            $locationRadiusSetting->setValue($request->request->get('locationRadiusMeters'));
-            $entityManager->persist($locationRadiusSetting);
-
-            $locationMeasurementSetting = $settingsRepository->getSettingByName('locationMeasurement');
-            if (!$locationMeasurementSetting) {
-                $locationMeasurementSetting = new Settings();
-                $locationMeasurementSetting->setName('locationMeasurement');
-            }
-            $locationMeasurementSetting->setValue($request->request->get('locationMeasurement'));
-            $entityManager->persist($locationMeasurementSetting);
-
-            $locationZoomSetting = $settingsRepository->getSettingByName('locationZoom');
-            if (!$locationZoomSetting) {
-                $locationZoomSetting = new Settings();
-                $locationZoomSetting->setName('locationZoom');
-            }
-            $locationZoomSetting->setValue($request->request->get('locationZoom'));
-            $entityManager->persist($locationZoomSetting);
-
-            $languageSetting = $settingsRepository->getSettingByName('language');
-            if (!$languageSetting) {
-                $languageSetting = new Settings();
-                $languageSetting->setName('language');
-            }
-            $languageSetting->setValue($request->request->get('language'));
-            $entityManager->persist($languageSetting);
-
-            $locationLatLngSetting = $settingsRepository->getSettingByName('locationLatLng');
-            if (!$locationLatLngSetting) {
-                $locationLatLngSetting = new Settings();
-                $locationLatLngSetting->setName('locationLatLng');
-            }
-            $locationLatLngSetting->setValue($request->request->get('locationLatLng'));
-            $entityManager->persist($locationLatLngSetting);
-
-            $timezoneSetting = $settingsRepository->getSettingByName('timezone');
-            if (!$timezoneSetting) {
-                $timezoneSetting = new Settings();
-                $timezoneSetting->setName('timezone');
-            }
+            $this->persistSetting('habitatName', $request->request->get('habitatName'));
+            $this->persistSetting('locationRadiusMeters', $request->request->get('locationRadiusMeters'));
+            $this->persistSetting('locationMeasurement', $request->request->get('locationMeasurement'));
+            $this->persistSetting('locationZoom', $request->request->get('locationZoom'));
+            $this->persistSetting('language', $request->request->get('language'));
+            $this->persistSetting('locationLatLng', $request->request->get('locationLatLng'));
             $timezone = LatLong::getTimezone($request->request->get('locationLatLng'));
-            $timezoneSetting->setValue($timezone->getName());
-            $entityManager->persist($timezoneSetting);
+            $this->persistSetting('timezone', $timezone->getName());
+            $this->persistSetting('theme', $request->request->get('theme'));
 
             $sidebarContent->setContent($request->request->get('sidebarContent'));
-            $entityManager->persist($sidebarContent);
-
-            $entityManager->flush();
+            $this->entityManager->persist($sidebarContent);
+            $this->entityManager->flush();
 
             $this->addFlash(
                 'notice',
@@ -137,20 +111,23 @@ class SettingsController extends AbstractController
             return $this->redirectToRoute('app_admin_settings');
         }
 
-        $habitatNameSetting = $settingsRepository->getSettingByName('habitatName');
-        $locationLatLngSetting = $settingsRepository->getSettingByName('locationLatLng');
-        $locationZoomSetting = $settingsRepository->getSettingByName('locationZoom');
-        $languageSetting = $settingsRepository->getSettingByName('language');
-        $locationMeasurementSetting = $settingsRepository->getSettingByName('locationMeasurement');
-        $locationRadiusSetting = $settingsRepository->getSettingByName('locationRadiusMeters');
+        $habitatNameSetting = $this->settingsRepository->getSettingByName('habitatName');
+        $locationLatLngSetting = $this->settingsRepository->getSettingByName('locationLatLng');
+        $locationZoomSetting = $this->settingsRepository->getSettingByName('locationZoom');
+        $languageSetting = $this->settingsRepository->getSettingByName('language');
+        $themeSetting = $this->settingsRepository->getSettingByName('theme');
+        $locationMeasurementSetting = $this->settingsRepository->getSettingByName('locationMeasurement');
+        $locationRadiusSetting = $this->settingsRepository->getSettingByName('locationRadiusMeters');
 
         return $this->render('admin/settings.html.twig', [
             'languages' => SetupController::LANGUAGES,
+            'adminUserTheme' => $adminUserTheme,
             'values' => [
                 'habitatName' => ($habitatNameSetting) ? $habitatNameSetting->getValue() : '',
                 'locationLatLng' => ($locationLatLngSetting) ? $locationLatLngSetting->getValue() : '51,0',
                 'locationZoom' => ($locationZoomSetting) ? $locationZoomSetting->getValue() : '3',
                 'language' => ($languageSetting) ? $languageSetting->getValue() : 'en',
+                'theme' => ($themeSetting) ? $themeSetting->getValue() : 'light',
                 'locationMeasurement' => ($locationMeasurementSetting) ? $locationMeasurementSetting->getValue() : 'km',
                 'locationRadiusMeters' => ($locationRadiusSetting) ? $locationRadiusSetting->getValue() : '3000',
                 'sidebarContent' => ($sidebarContent->getContent()) ? $sidebarContent->getContent() : '',
@@ -194,6 +171,25 @@ class SettingsController extends AbstractController
             $errors['language'][] = $this->translator->trans('fields.language.validations.empty');
         }
 
+        if (
+            empty($request->request->get('theme'))
+            || !in_array($request->request->get('theme'), ['light', 'dark'])
+        ) {
+            $errors['theme'][] = $this->translator->trans('fields.theme.validations.empty');
+        }
+
         return $errors;
+    }
+
+    private function persistSetting(string $settingName, string $settingValue): void
+    {
+        $setting = $this->settingsRepository->getSettingByName($settingName);
+        if (!$setting) {
+            $setting = new Settings();
+            $setting->setName($settingName);
+        }
+        $setting->setValue($settingValue);
+
+        $this->entityManager->persist($setting);
     }
 }