+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 
+- [#123] Dark mode
 - [#131] A message displays when an anonymous user clicks a heart
+
+### Fixed
+
+- Do not expect only database ID to be 1 for any entity
 
 ## [1.4.0] - 2026-05-16